Generalize responsibles between Serviceform items + Serviceform
