+++ serviceform/serviceform/models/__init__.py
@@ -21,7 +21,6 @@ from .participation import (ParticipationActivity, ParticipationActivityChoice, 
                             QuestionAnswer, Participation)
 from .people import Organization, Member
 from .serviceform import (ServiceForm, FormRevision, Activity, ActivityChoice, Level1Category,
-                          Level2Category, Question, ColorField)
+                          Level2Category, Question)
 
-
-from .mixins import (CopyMixin, NameDescriptionMixin, SubitemMixin)
+from .mixins import NameDescriptionMixin
